Prevent header from hiding top of section on jump to target anchor
The theme provides anchors to jump directly to a given section on a page
(From URL or from the "Permalink to this headline" buttons). When
jumping with it, the top of the target section is hidden under the
header banner.

Fix it by adding a hidden block of the same height of the .wy-header
class elements (70px + 1px padding) before the targets. Hide it when the
page is less than 790px wide (since the header is hidden too).

Signed-off-by: Quentin Monnet <quentin@isovalent.com>