Fix typo on .rst-current-version-label class name
When rebasing from branch v0.7 on top of upstream to create branch v1.0,
a typo was introduced on a class name. As a result, the "Read the Docs"
label in the bottom-left corner of the HTML pages are not styled
properly (it misses a heavier font and it should float to the left).
Let's fix the typo.

Fixes: 3eb695b4444c ("Initial commit for Cilium theme")
Signed-off-by: Quentin Monnet <quentin@isovalent.com>